logout

+++ client/src/pages/Auth/Logout.jsx
@@ -1,33 +1,32 @@
 import { Dialog, DialogBody, DialogHeader, DialogFooter } from "@material-tailwind/react";
 import { Button, Card, CardHeader, Typography, CardFooter, CardBody, Input, Select, Option } from "@material-tailwind/react";
-
 import { useNavigate } from "react-router-dom";
+
+import { useLogoutAdminMutation } from "../../redux/services/authApi";
+import { useEffect } from "react";
+import { toast } from "react-toastify";
 
 const Logout = ({ close }) => {
     const navigate = useNavigate()
 
-    const onLogout = () => {
-        localStorage.clear()
-        navigate('/login')
+    const [logout, responseLogout] = useLogoutAdminMutation();
+
+    useEffect(() => {
+        if (responseLogout.isSuccess) {
+            localStorage.clear()
+            navigate('/login')
+        } else if (responseLogout.isError) {
+            toast.dismiss()
+            toast.error('Something wrong, please contact the admin');
+        }
+    }, [responseLogout])
+
+    const onLogout = async () => {
+        await logout()
     }
     return (
         <>
             <Dialog className="bg-transparent shadow-none" size="xs" open={true}>
-                {/* <DialogHeader>Logout</DialogHeader>
-                <DialogBody divider style={{ fontSize: 20, fontWeight: 'bold' }}>
-                    Are you sure?
-                </DialogBody>
-                <DialogFooter>
-                    <div className="flex justify-end gap-2">
-                        <Button className="bg-gray-800 hover:bg-red-400" size="md" onClick={close}>
-                            Cancel
-                        </Button>
-                        <Button className="bg-black hover:bg-blue-600" size="md" type="button" onClick={onLogout}>
-                            Confirm
-                        </Button>
-                    </div>
-                </DialogFooter> */}
-
                 <Card className="mx-auto w-full max-w-[24rem]">
                     <CardHeader
                         color="gray"
